moved TranslationManager class to a separate file
